+++ src/compas_timber/connections/generic_joint.py
@@ -50,7 +50,7 @@ class GenericJoint(Joint):
         self.element_b = element_b
         self.element_a_guid = str(element_a.guid) if element_a else None
         self.element_b_guid = str(element_b.guid) if element_b else None
-        self.distance = distance or TOL.absolute
+        self.distance = distance if distance is not None else None
 
     @property
     def elements(self):