+++ src/compas_timber/design/workflow.py
@@ -35,6 +35,49 @@ class ContainerDefinition(object):
         self.config_set = config_set
 
 
+class JointRuleSolver(object):
+    """A class that holds a set of rules for joining elements.
+
+    It can be used to apply the rules to a model and create joints based on the rules.
+
+    Parameters
+    ----------
+    rules : list(:class:`~compas_timber.connections.JointRule`)
+        A list of rules to apply to the model.
+    """
+
+    def __init__(self, rules, use_defaults=False):
+        self.rules = rules
+        self.use_defaults = use_defaults
+
+
+    def apply(self, model, max_distance=TOL.absolute, handled_pairs=None):
+        return JointRule.apply_joint_rules_to_model(self.rules, model, max_distance, handled_pairs)
+    
+    @staticmethod
+    def direct_rules(rules):
+        return [rule for rule in rules if rule.__class__.__name__ == "DirectRule"]
+
+    @staticmethod
+    def category_rules(rules):
+        return [rule for rule in rules if rule.__class__.__name__ == "CategoryRule"]
+
+    @staticmethod
+    def topology_rules(rules, use_defaults=False):
+        topo_rules = {}
+        if use_defaults:
+            topo_rules = {
+                JointTopology.TOPO_L: TopologyRule(JointTopology.TOPO_L, LMiterJoint),
+                JointTopology.TOPO_T: TopologyRule(JointTopology.TOPO_T, TButtJoint),
+                JointTopology.TOPO_X: TopologyRule(JointTopology.TOPO_X, XLapJoint),
+                JointTopology.TOPO_EDGE_EDGE: TopologyRule(JointTopology.TOPO_EDGE_EDGE, PlateMiterJoint),
+                JointTopology.TOPO_EDGE_FACE: TopologyRule(JointTopology.TOPO_EDGE_FACE, PlateTButtJoint),
+            }
+        for rule in rules:  # separate category and topo and direct joint rules
+            if rule.__class__.__name__ == "TopologyRule":
+                topo_rules[rule.topology_type] = TopologyRule(rule.topology_type, rule.joint_type, rule.max_distance, **rule.kwargs)  # overwrites, meaning last rule wins
+        return [rule for rule in topo_rules.values() if rule is not None]
+
 class JointRule(object):
     def __init__(self, joint_type, max_distance=None, **kwargs):
         """Initializes a JointRule with the given joint type and optional max distance."""
@@ -67,7 +110,7 @@ class JointRule(object):
         return [rule for rule in topo_rules.values() if rule is not None]
 
     @staticmethod
-    def joints_from_rules_and_model(rules, model, max_distance=TOL.absolute, handled_pairs=None):
+    def apply_joint_rules_to_model(rules, model, max_distance=TOL.absolute, handled_pairs=None):
         """Returns a list of joints based on the given rules and elements.
 
         Parameters
@@ -91,7 +134,8 @@ class JointRule(object):
         handled_pairs = handled_pairs or []
 
         max_rule_distance = max([rule.max_distance for rule in rules if rule.max_distance] + [max_distance])
-
+        print("max_rule_distance, ",max_rule_distance)
+        
         model.connect_adjacent_beams(max_distance=max_rule_distance)  # ensure that the model is connected before analyzing
         model.connect_adjacent_plates(max_distance=max_rule_distance)  # ensure that the model is connected before analyzing
         analyzer = MaxNCompositeAnalyzer(model, n=len(list(model.elements())))
@@ -101,17 +145,12 @@ class JointRule(object):
         # however, other beams in the model should be allowed to be joined with them, thus they cannot be altogether excluded
         clusters_temp = [c for c in clusters]
         for cluster in clusters_temp:
-            print([j.distance for j in cluster.joints])
             if set(cluster.elements) in handled_pairs:
                 clusters.remove(cluster)
-        print("len(clusters)", len(clusters))
         joining_errors = []
         joining_errors.extend(JointRule.process_rules(JointRule.get_direct_rules(rules), model, clusters, max_distance=max_distance))
         joining_errors.extend(JointRule.process_rules(JointRule.get_category_rules(rules), model, clusters, max_distance=max_distance))
-        print("len(clusters)", len(clusters))
         joining_errors.extend(JointRule.process_rules(JointRule.get_topology_rules(rules, use_defaults=False), model, clusters, max_distance=max_distance))
-        print("len(clusters)", len(clusters))
-        print([j.__class__.__name__ for j in model.joints])
         return joining_errors, clusters
 
     def _comply_topology(self, cluster, raise_error=False):
@@ -143,10 +182,11 @@ class JointRule(object):
         return True
 
     def _comply_distance(self, cluster, max_distance=None, raise_error=False):
-        max_distance = self.max_distance or max_distance
+        max_distance = self.max_distance or max_distance or None
         if not max_distance:
             return True
         distance = max([j.distance for j in cluster.joints])
+        print("j.dist",[j.distance for j in cluster.joints])
         if distance > max_distance:
             if raise_error:
                 raise BeamJoiningError(
@@ -177,7 +217,6 @@ class JointRule(object):
         joining_errors = []
         for rule in rules:
             for cluster in [c for c in clusters]:
-                print([j.distance for j in cluster.joints])
                 joint, error = rule.try_create_joint(model, cluster, max_distance=max_distance)
                 if joint:
                     clusters.remove(cluster)  # remove the cluster from the list of clusters to avoid processing it again