+++ tests/compas_timber/test_joint_rules.py
@@ -103,6 +103,12 @@ def L_beams_separated():
     ]
     return [Beam.from_centerline(line, w, h) for line in lines]
 
+@pytest.fixture
+def model(beams):
+    model = TimberModel()
+    model.add_elements(beams)
+    return model
+
 
 def test_joints_from_beams_and_topo_rules(beams):
     model = TimberModel()
@@ -112,12 +118,12 @@ def test_joints_from_beams_and_topo_rules(beams):
         TopologyRule(JointTopology.TOPO_T, TButtJoint),
         TopologyRule(JointTopology.TOPO_X, XLapJoint),
     ]
-    errors, unjoined_clusters = JointRule.joints_from_rules_and_model(rules, model)
+    errors, unjoined_clusters = JointRule.apply_joint_rules_to_model(rules, model)
     assert len(model.joints) == 4
     assert set([joint.__class__.__name__ for joint in model.joints]) == set(["LMiterJoint", "TButtJoint", "XLapJoint"])
 
 
-def test_joints_from_beams_and_rules_with_max_distance(separated_beams):
+def test_joints_from_beams_and_rules_with_no_max_distance(separated_beams):
     model = TimberModel()
     model.add_elements(separated_beams)
     rules = [
@@ -126,9 +132,10 @@ def test_joints_from_beams_and_rules_with_max_distance(separated_beams):
         TopologyRule(JointTopology.TOPO_X, XLapJoint),
     ]
     with pytest.raises(ValueError):
-        errors, unjoined_clusters = JointRule.joints_from_rules_and_model(rules, model)
+        errors, unjoined_clusters = JointRule.apply_joint_rules_to_model(rules, model)
     assert len(model.joints) == 0
 
+def test_joints_from_beams_and_rules_with_max_distance_rule(separated_beams):
     model = TimberModel()
     model.add_elements(separated_beams)
     rules = [
@@ -136,10 +143,11 @@ def test_joints_from_beams_and_rules_with_max_distance(separated_beams):
         TopologyRule(JointTopology.TOPO_T, TButtJoint, max_distance=0.15),
         TopologyRule(JointTopology.TOPO_X, XLapJoint),
     ]
-    errors, unjoined_clusters = JointRule.joints_from_rules_and_model(rules, model)
+    errors, unjoined_clusters = JointRule.apply_joint_rules_to_model(rules, model)
     assert len(model.joints) == 4
     assert len(unjoined_clusters) == 3
 
+def test_joints_from_beams_and_rules_with_max_distance_model(separated_beams):
     model = TimberModel()
     model.add_elements(separated_beams)
     rules = [
@@ -147,7 +155,7 @@ def test_joints_from_beams_and_rules_with_max_distance(separated_beams):
         TopologyRule(JointTopology.TOPO_T, TButtJoint, max_distance=0.05),
         TopologyRule(JointTopology.TOPO_X, XLapJoint),
     ]
-    errors, unjoined_clusters = JointRule.joints_from_rules_and_model(rules, model, max_distance=0.15)
+    errors, unjoined_clusters = JointRule.apply_joint_rules_to_model(rules, model, max_distance=0.15)
     assert len(model.joints) == 4
     assert len(unjoined_clusters) == 1
 
@@ -160,31 +168,35 @@ def test_direct_rule_matches_cluster(beams):
 
 
 def test_direct_rule_get_joint(beams):
-    rule = DirectRule(LMiterJoint, [beams[0], beams[1]])
-    assert rule.get_joint() is not None
-    rule = DirectRule(LMiterJoint, [beams[2], beams[3]])
-    assert rule.get_joint() is not None
-    rule = DirectRule(TButtJoint, [beams[2], beams[1]])
-    assert rule.get_joint() is not None
-    rule = DirectRule(XLapJoint, [beams[3], beams[0]])
-    assert rule.get_joint() is not None
+    model = TimberModel()
+    model.add_elements(beams)
+    rules = [DirectRule(LMiterJoint, [beams[0], beams[1]]),
+        DirectRule(LMiterJoint, [beams[2], beams[3]]),
+        DirectRule(TButtJoint, [beams[2], beams[1]]),
+        DirectRule(XLapJoint, [beams[3], beams[0]]),]
+    JointRule.apply_joint_rules_to_model(rules, model)
+    assert set([j.__class__.__name for j in model.joints]) == set(["LMiterJoint","TButtJoint","XLapJoint"])
 
 
 def test_direct_rule_get_joint_max_distance(separated_beams):
+    model = TimberModel()
+    model.add_elements(separated_beams)
     rule = DirectRule(LMiterJoint, [separated_beams[0], separated_beams[1]], max_distance=0.05)
     with pytest.raises(BeamJoiningError):
-        rule.get_joint()
+        rule.apply_joint_rules_to_model([rule], model)
     rule = DirectRule(LMiterJoint, [separated_beams[0], separated_beams[1]], max_distance=0.15)
-    assert rule.get_joint() is not None
+    assert rule.apply_joint_rules_to_model() is not None
 
 
 def test_category_rule_try_get_joint(beams):
+    model = TimberModel()
+    model.add_elements(separated_beams)
     for beam in beams:
         beam.attributes["category"] = "A"
     beams[1].attributes["category"] = "B"
     rule = CategoryRule(LMiterJoint, "A", "B")
-    assert rule.try_get_joint(beams[:2]) is not None
-    assert rule.try_get_joint(beams[2:]) is None
+    assert JointRule.apply_joint_rules_to_model([rule], model) is not None
+    assert JointRule.apply_joint_rules_to_model(beams[2:], model) is None
 
 
 def test_topology_rule_try_get_joint(beams):
@@ -200,7 +212,7 @@ def test_different_rules(L_beams):
         beam.attributes["category"] = "A"
     L_beams[1].attributes["category"] = "B"
     rules = [DirectRule(LLapJoint, L_beams[:2]), CategoryRule(LButtJoint, "A", "B"), TopologyRule(JointTopology.TOPO_L, LMiterJoint)]
-    joints, errors = JointRule.joints_from_rules_and_model(rules, L_beams)
+    joints, errors = JointRule.apply_joint_rules_to_model(rules, L_beams)
     assert len(joints) == 3
     assert set([joint.__class__.__name__ for joint in joints]) == set(["LLapJoint", "LButtJoint", "LMiterJoint"])
 
@@ -210,12 +222,12 @@ def test_different_rules_max_distance(L_beams_separated):
         beam.attributes["category"] = "A"
     L_beams_separated[1].attributes["category"] = "B"
     rules = [DirectRule(LLapJoint, L_beams_separated[:2]), CategoryRule(LButtJoint, "A", "B"), TopologyRule(JointTopology.TOPO_L, LMiterJoint)]
-    joints, errors = JointRule.joints_from_rules_and_model(rules, L_beams_separated)
+    joints, errors = JointRule.apply_joint_rules_to_model(rules, L_beams_separated)
     assert len(joints) == 0
 
     rules = [DirectRule(LLapJoint, L_beams_separated[:2]), CategoryRule(LButtJoint, "A", "B"), TopologyRule(JointTopology.TOPO_L, LMiterJoint, max_distance=0.15)]
 
-    joints, errors = JointRule.joints_from_rules_and_model(rules, L_beams_separated)
+    joints, errors = JointRule.apply_joint_rules_to_model(rules, L_beams_separated)
     assert len(joints) == 3
     assert set([joint.__class__.__name__ for joint in joints]) == set(["LMiterJoint"])
 
@@ -237,7 +249,7 @@ def test_plate_topo_rules():
         TopologyRule(JointTopology.TOPO_EDGE_FACE, PlateTButtJoint),
         TopologyRule(JointTopology.TOPO_EDGE_EDGE, PlateMiterJoint),
     ]
-    joints, errors = JointRule.joints_from_rules_and_model(rules, [plate_a, plate_b, plate_c])
+    joints, errors = JointRule.apply_joint_rules_to_model(rules, [plate_a, plate_b, plate_c])
     assert len(joints) == 3, "Expected three joints"
 
 
@@ -259,7 +271,7 @@ def test_plate_category_rules():
         CategoryRule(PlateMiterJoint, "A", "C"),
         CategoryRule(PlateLButtJoint, "B", "C"),
     ]
-    joints, errors = JointRule.joints_from_rules_and_model(rules, [plate_a, plate_b, plate_c])
+    joints, errors = JointRule.apply_joint_rules_to_model(rules, [plate_a, plate_b, plate_c])
     assert len(joints) == 2, "Expected two joints"
 
     plate_a.reset()
@@ -271,7 +283,7 @@ def test_plate_category_rules():
         CategoryRule(PlateMiterJoint, "A", "C"),
         CategoryRule(PlateLButtJoint, "B", "C"),
     ]
-    joints, errors = JointRule.joints_from_rules_and_model(rules, [plate_a, plate_b, plate_c])
+    joints, errors = JointRule.apply_joint_rules_to_model(rules, [plate_a, plate_b, plate_c])
 
 
 def test_plate_rules_priority():
@@ -295,7 +307,7 @@ def test_plate_rules_priority():
         TopologyRule(JointTopology.TOPO_EDGE_FACE, PlateTButtJoint),
         TopologyRule(JointTopology.TOPO_EDGE_EDGE, PlateMiterJoint),
     ]
-    joints, errors = JointRule.joints_from_rules_and_model(rules, [plate_a, plate_b, plate_c])
+    joints, errors = JointRule.apply_joint_rules_to_model(rules, [plate_a, plate_b, plate_c])
     assert len(joints) == 3, "Expected three joints"
     assert set([j.__class__.__name__ for j in joints]) == set(["PlateLButtJoint", "PlateTButtJoint", "PlateMiterJoint"]), (
         "Expected PlateLButtJoint, PlateTButtJoint, and PlateMiterJoint"
@@ -308,7 +320,7 @@ def test_plate_rules_priority():
         TopologyRule(JointTopology.TOPO_EDGE_FACE, PlateTButtJoint),
         TopologyRule(JointTopology.TOPO_EDGE_EDGE, PlateMiterJoint),
     ]
-    joints, errors = JointRule.joints_from_rules_and_model(rules, [plate_a, plate_b, plate_c])
+    joints, errors = JointRule.apply_joint_rules_to_model(rules, [plate_a, plate_b, plate_c])
     assert len(joints) == 3, "Expected three joints"
     assert set([j.__class__.__name__ for j in joints]) == set(["PlateLButtJoint", "PlateTButtJoint"]), "Expected PlateLButtJoint, PlateTButtJoint, and PlateMiterJoint"
 
@@ -318,7 +330,7 @@ def test_plate_rules_priority():
         TopologyRule(JointTopology.TOPO_EDGE_FACE, PlateTButtJoint),
         TopologyRule(JointTopology.TOPO_EDGE_EDGE, PlateMiterJoint),
     ]
-    joints, errors = JointRule.joints_from_rules_and_model(rules, [plate_a, plate_b, plate_c])
+    joints, errors = JointRule.apply_joint_rules_to_model(rules, [plate_a, plate_b, plate_c])
     assert len(joints) == 3, "Expected three joints"
     assert set([j.__class__.__name__ for j in joints]) == set(["PlateLButtJoint", "PlateTButtJoint", "PlateMiterJoint"]), (
         "Expected PlateLButtJoint, PlateTButtJoint, and PlateMiterJoint"